ai_model/requirements.txt
pandas>=1.5.0
matplotlib>=3.5.0
Pillow>=9.0.0
jupyter>=1.0.0

